UPD: OpenMaster name & behaviour on loading project
